+++ pkg/web/static/preline/preline.js.LICENSE.txt
@@ -1,0 +1,81 @@
+/*
+ * Component
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSAccordion
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSCollapse
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSDropdown
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @requires: @popperjs/core ^2.11.2
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSOverlay
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Htmlstream Libraries (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSRemoveElement
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSScrollspy
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * HSTabs
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+ * Util
+ * @version: 1.3.0
+ * @author: HtmlStream
+ * @license: Licensed under MIT (https://preline.co/docs/license.html)
+ * Copyright 2023 Htmlstream
+ */
+
+/*
+* HSTooltips
+* @version: 1.4.0
+* @author: HtmlStream
+* @requires: @popperjs/core ^2.11.2
+* @license: Licensed under MIT (https://preline.co/docs/license.html)
+* Copyright 2023 Htmlstream
+*/
