+++ pkg/web/templates/components/chat_history.html
@@ -22,7 +22,7 @@
                             </ul>
                         </div>
                     </div>
-                    {{ template "MessagePager" . }}
+                    {{ template "PageCountPager" . }}
                 </div>
             </div>
         </div>
