+++ pkg/web/templates/components/layout/scripts_top.html
@@ -1,4 +1,4 @@
 {{define "ScriptsTop"}}
 <script type="module" src="/static/js/htmx.min.js"></script>
-<script defer src="/static/alpinejs-3.13.0.min.js"></script>
+<script defer src="/static/js/alpinejs-3.13.0.min.js"></script>
 {{end}}
