+++ pkg/web/templates/components/session_table.html
@@ -139,7 +139,7 @@
                     <!-- End Table -->
 
                     <!-- Footer -->
-                    {{ template "TablePager" . }}
+                    {{ template "CursorPager" . }}
                     <!-- End Footer -->
                     {{ end }}
                 </div>
