+++ pkg/web/templates/pages/user_details.html
@@ -23,7 +23,8 @@
 <div class="px-4 py-5 sm:px-6 lg:px-8 lg:py-7 mx-auto">
     <!-- Card -->
     <div class="bg-white rounded-xl shadow p-4 sm:p-7 dark:bg-slate-900">
-        <button type="button"
+        <button type="button" 
+            hx-confirm="Are you sure you want to delete this user?" hx-delete="{{ .Path }}" hx-target="#page-content"
             class="py-3 px-4 inline-flex justify-center items-center gap-2 rounded-md bg-red-100 border border-transparent font-semibold text-red-500 hover:text-white hover:bg-red-500 focus:outline-none focus:ring-2 ring-offset-white focus:ring-red-500 focus:ring-offset-2 transition-all text-sm dark:focus:ring-offset-gray-800">
             Delete User {{ .Data.UserID }}
         </button>
